moved location arrays into enum file renamed locationTypes
